requirements.txt
django
celery
redis
djangorestframework
django-cors-headers
drf-spectacular